Move /balances and /chains to /routes (#77)
