+++ src/routes/chains/chains.service.ts
@@ -2,12 +2,12 @@ import { Inject, Injectable } from '@nestjs/common';
 import {
   cursorUrlFromLimitAndOffset,
   PaginationData,
-} from '../common/pagination/pagination.data';
-import { IChainsRepository } from '../domain/chains/chains.repository.interface';
-import { IBackboneRepository } from '../domain/backbone/backbone.repository.interface';
-import { Page } from '../domain/entities/page.entity';
-import { Chain } from '../domain/chains/entities/chain.entity';
-import { Backbone } from '../domain/backbone/entities/backbone.entity';
+} from '../../common/pagination/pagination.data';
+import { IChainsRepository } from '../../domain/chains/chains.repository.interface';
+import { IBackboneRepository } from '../../domain/backbone/backbone.repository.interface';
+import { Page } from '../../domain/entities/page.entity';
+import { Chain } from '../../domain/chains/entities/chain.entity';
+import { Backbone } from '../../domain/backbone/entities/backbone.entity';
 
 @Injectable()
 export class ChainsService {
